+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@ Release Notes
 ## 3.14.0
 
 - Fixed RSS feed date time strings to format correctly according to the spec.
+- Added `atom` package for Atom feeds.
 
 ## 3.13.0